+++ app/routes.js
@@ -4,13 +4,13 @@ import { Route, IndexRoute } from 'react-router';
 import MainLayout from './layouts/MainLayout';
 import NotFound from './pages/NotFound';
 import Home from './pages/Home';
-import Timer from './components/Timer';
+import Settings from './pages/Settings';
 
 
 export default (
   <Route path="/" component={MainLayout}>
     <IndexRoute component={Home} />
-    <Route path="/timer" component={Timer} />
+    <Route path="/settings" component={Settings} />
     <Route path="*" component={NotFound} />
   </Route>
 );